Reorganize sigchain verify to reduce go sdk memory usage on mobile
