+++ pb/sigchain.go
@@ -10,9 +10,6 @@ import (
 
 	"github.com/nknorg/nkn/v2/common"
 	"github.com/nknorg/nkn/v2/common/serialization"
-	"github.com/nknorg/nkn/v2/crypto"
-	"github.com/nknorg/nkn/v2/config"
-	"github.com/nknorg/nnet/overlay/chord"
 )
 
 const (
@@ -137,159 +134,6 @@ func (sc *SigChain) SerializationMetadata(w io.Writer) error {
 	err = serialization.WriteVarBytes(w, sc.DestPubkey)
 	if err != nil {
 		return err
-	}
-
-	return nil
-}
-
-func (sc *SigChain) Verify(height uint32) error {
-	if err := sc.VerifyMeta(height); err != nil {
-		return err
-	}
-
-	if err := sc.VerifyPath(); err != nil {
-		return err
-	}
-
-	if err := sc.VerifySignatures(); err != nil {
-		return err
-	}
-
-	return nil
-}
-
-func (sc *SigChain) VerifyMeta(height uint32) error {
-	if sc.Length() < 3 {
-		return fmt.Errorf("sigchain should have at least 3 elements, but only has %d", sc.Length())
-	}
-
-	if len(sc.Elems[0].Id) > 0 && !bytes.Equal(sc.SrcId, sc.Elems[0].Id) {
-		return fmt.Errorf("sigchain has wrong src id")
-	}
-
-	if !sc.IsComplete() {
-		return fmt.Errorf("sigchain is not complete")
-	}
-
-	if bytes.Equal(sc.Elems[0].Id, sc.Elems[1].Id) {
-		return fmt.Errorf("src and first relayer has the same ID")
-	}
-
-	if bytes.Equal(sc.Elems[sc.Length()-1].Id, sc.Elems[sc.Length()-2].Id) {
-		return fmt.Errorf("dest and last relayer has the same ID")
-	}
-
-	if len(sc.Elems[sc.Length()-1].NextPubkey) > 0 {
-		return fmt.Errorf("next pubkey in last sigchain elem should be empty")
-	}
-
-	if sc.Elems[0].Mining {
-		return fmt.Errorf("first sigchain element should have mining set to false")
-	}
-
-	if sc.Elems[sc.Length()-1].Mining {
-		return fmt.Errorf("last sigchain element should have mining set to false")
-	}
-
-	for i, e := range sc.Elems {
-		if i == 0 || i == sc.Length()-1 {
-			if len(e.Vrf) > 0 {
-				return fmt.Errorf("sigchain elem %d vrf should be empty", i)
-			}
-			if len(e.Proof) > 0 {
-				return fmt.Errorf("sigchain elem %d proof should be empty", i)
-			}
-		} else {
-			if len(e.Vrf) == 0 {
-				return fmt.Errorf("sigchain elem %d vrf should not be empty", i)
-			}
-			if len(e.Proof) == 0 {
-				return fmt.Errorf("sigchain elem %d proof should not be empty", i)
-			}
-		}
-		if !config.AllowSigChainHashSignature.GetValueAtHeight(height) {
-			if e.SigAlgo != SIGNATURE {
-				return fmt.Errorf("sigchain elem %d sig algo should be %v", i, SIGNATURE)
-			}
-		}
-	}
-
-	return nil
-}
-
-// VerifySignatures returns whether all signatures in sigchain are valid
-func (sc *SigChain) VerifySignatures() error {
-	prevNextPubkey := sc.SrcPubkey
-	buff := bytes.NewBuffer(nil)
-	err := sc.SerializationMetadata(buff)
-	if err != nil {
-		return err
-	}
-	metaHash := sha256.Sum256(buff.Bytes())
-	prevHash := metaHash[:]
-	for i, e := range sc.Elems {
-		err := crypto.CheckPublicKey(prevNextPubkey)
-		if err != nil {
-			return fmt.Errorf("invalid pubkey %x: %v", prevNextPubkey, err)
-		}
-
-		if sc.IsComplete() && i == sc.Length()-1 {
-			h := sha256.Sum256(prevHash)
-			prevHash = h[:]
-		}
-
-		hash, err := e.Hash(prevHash)
-		if err != nil {
-			return err
-		}
-
-		switch e.SigAlgo {
-		case SIGNATURE:
-			err = crypto.Verify(prevNextPubkey, hash, e.Signature)
-			if err != nil {
-				return fmt.Errorf("signature %x is invalid: %v", e.Signature, err)
-			}
-		case HASH:
-			if !bytes.Equal(e.Signature, hash) {
-				return fmt.Errorf("signature %x is different from expected value %x", e.Signature, hash[:])
-			}
-		default:
-			return fmt.Errorf("unknown SigAlgo %v", e.SigAlgo)
-		}
-
-		if len(e.Vrf) > 0 {
-			ok := crypto.VerifyVrf(prevNextPubkey, sc.BlockHash, e.Vrf, e.Proof)
-			if !ok {
-				return fmt.Errorf("invalid vrf or proof")
-			}
-			prevHash = hash
-		} else {
-			prevHash = e.Signature
-		}
-
-		prevNextPubkey = e.NextPubkey
-		if sc.IsComplete() && i == sc.Length()-2 && len(e.NextPubkey) == 0 {
-			prevNextPubkey = sc.DestPubkey
-		}
-	}
-
-	return nil
-}
-
-func (sc *SigChain) VerifyPath() error {
-	var t big.Int
-	lastNodeID := sc.Elems[sc.Length()-2].Id
-	prevDistance := chord.Distance(sc.Elems[1].Id, lastNodeID, config.NodeIDBytes*8)
-	for i := 2; i < sc.Length()-1; i++ {
-		dist := chord.Distance(sc.Elems[i].Id, lastNodeID, config.NodeIDBytes*8)
-		if dist.Cmp(prevDistance) == 0 {
-			return fmt.Errorf("relayer %d and %d has the same ID", i-1, i)
-		}
-		(&t).Mul(dist, big.NewInt(2))
-		if t.Cmp(prevDistance) > 0 {
-			return fmt.Errorf("signature chain path is invalid")
-		}
-		prevDistance = dist
 	}
 
 	return nil
@@ -468,4 +312,16 @@ func ComputeSignatureHash(lastRelayHash []byte, sigChainLen int) []byte {
 	sigHash := h[:]
 	rightShiftBytes(sigHash, bitShiftPerSigChainElement*sigChainLen)
 	return sigHash
+}
+
+func rightShiftBytes(b []byte, n int) {
+	for k := 0; k < n; k++ {
+		b[len(b)-1] >>= 1
+		for i := len(b) - 2; i >= 0; i-- {
+			if b[i]&0x1 == 0x1 {
+				b[i+1] |= 0x80
+			}
+			b[i] >>= 1
+		}
+	}
 }
